Move   `GalleryViewController` to production
