--- NASAGallery/NASAGalleryiOS/SampleFile.swift
@@ -1,8 +1,0 @@
-//
-//  SampleFile.swift
-//  NASAGalleryiOS
-//
-//  Created by Ivo on 27/03/25.
-//
-
-import Foundation
